+++ src/main/java/com/test/webtest/domain/webvitals/sub/dto/ClsDto.java
@@ -1,4 +1,4 @@
-package com.test.webtest.domain.webvitals.ai.dto;
+package com.test.webtest.domain.webvitals.sub.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
