+++ src/main/java/com/test/webtest/domain/webvitals/sub/dto/FcpDto.java
@@ -1,4 +1,4 @@
-package com.test.webtest.domain.webvitals.ai.dto;
+package com.test.webtest.domain.webvitals.sub.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -8,35 +8,24 @@ import java.util.List;
 
 @Getter
 @NoArgsConstructor
-public class InpDto {
+public class FcpDto {
     private String name;
     private Double value;
     private String rating;
-    private List<InpEntry> entries;
+    private List<FcpEntry> entries;
 
     @Getter
     @NoArgsConstructor
-    public static class InpEntry {
+    public static class FcpEntry {
         private String name;
 
         @JsonProperty("entryType")
         private String entryType;
 
         @JsonProperty("startTime")
-        private Double startTime;
+        private Integer startTime;
 
         private Integer duration;
-
-        @JsonProperty("interactionId")
-        private Integer interactionId;
-
-        @JsonProperty("processingStart")
-        private Double processingStart;
-
-        @JsonProperty("processingEnd")
-        private Double processingEnd;
-
-        private Boolean cancelable;
     }
 }
