+++ src/main/java/com/test/webtest/domain/webvitals/sub/dto/LcpDto.java
@@ -1,4 +1,4 @@
-package com.test.webtest.domain.webvitals.ai.dto;
+package com.test.webtest.domain.webvitals.sub.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -8,15 +8,15 @@ import java.util.List;
 
 @Getter
 @NoArgsConstructor
-public class FcpDto {
+public class LcpDto {
     private String name;
     private Double value;
     private String rating;
-    private List<FcpEntry> entries;
+    private List<LcpEntry> entries;
 
     @Getter
     @NoArgsConstructor
-    public static class FcpEntry {
+    public static class LcpEntry {
         private String name;
 
         @JsonProperty("entryType")
@@ -26,6 +26,18 @@ public class FcpDto {
         private Integer startTime;
 
         private Integer duration;
+
+        private Integer size;
+
+        @JsonProperty("renderTime")
+        private Integer renderTime;
+
+        @JsonProperty("loadTime")
+        private Integer loadTime;
+
+        private String id;
+
+        private String url;
     }
 }
